Add new test file, some test stubs, halfway done key class
I was going to finish writing the test stubs for the basic functions as well as add a copy constructor (to pass a student object into and initialize relevant studentKey fields from it) as well as a hash function for the student key. I think once those are in, were close to/ ready to write the functions themselves.